+++ packages.el
@@ -46,6 +46,7 @@
     org-super-agenda
     peg
     (org-ql :location local)
+    org-special-block-extras
 
     git-commit
     git-timemachine
@@ -689,6 +690,7 @@ which require an initialization must be listed explicitly in the list.")
 (defun eab-spacemacs/init-ov nil)
 (defun eab-spacemacs/init-peg nil)
 (defun eab-spacemacs/init-org-super-agenda nil)
+(defun eab-spacemacs/init-org-special-block-extras nil)
 (defun eab-spacemacs/init-org-ql nil
   (require 'org-ql)
   (require 'org-ql-search)